requirements.txt
Flask
Flask-Cors
Flask-SocketIO
Jinja2
MarkupSafe
Werkzeug
attrs
base58
boto3
botocore
cachetools
certifi
charset-normalizer
click
colorama
Flask-Session
future
greenlet
idna
itsdangerous
Jinja2-time
jmespath
jsonschema
MarkupSafe
numpy
openai
pandas
protobuf
pycryptodomex
python-dateutil
pytz
regex
request
s3transfer
six
soupsieve
tenacity
tqdm
typing-extensions
urllib3
uuid
Werkzeug
